5 files in Tomcat left!
That's right, I've massaged and kneaded COMTOR.g to the point where it
consumes almost all of the Tomcat source without any problems! Now I
just need to get those last 5 (which are all non-fatal errors) and move
on to more source code.
Eventually, I plan to have a suite of test programs (or test
repositories!) that I can just check any changes against. This will be
helpful for future optimisiations.

Most annoying bug: An empty if-statement body. It was empty because of
TODO!