Many tests passed
1) Added a directory for test code. Mostly source code from apache
projects.

2) Fixed many, many things in COMTOR.g. It has consumed over 7000 Java
source files without issue. (~5 files to give problems, but they're
broken on purpose!)

3) JavaComments is now more elegant and efficient. Instead of defining
all the things it *shouldn't* try to parse, such as Strings, which lead
to chars, which lead to whitespace and unicode, etc., it uses a gated
syntact predicate to determine if it's inside a string or not. If so, it
ignores comments. If not, it'll parse comments normally.